backend/pyproject.toml
[project]
name = "spiritisland"
version = "0.1.0"
description = ""
authors = [
    {name = "Duncan Sorenson",email = "Duncan@zagaran.com"}
]
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
isort = "^7.0.0"

